autotest-ent-360-brain/requirements.txt
robotframework
beautifulsoup4
requests
mysql-connector-python
robotframework-sshlibrary
PyYAML
xlrd